+++ packages/basehub/src/next/image/index.ts
@@ -1,1 +1,2 @@
 export { BaseHubImage, basehubImageLoader } from "./primitive";
+export type { BaseHubImageProps } from "./primitive";